Expanded Graph.cs functionality
Graph class now has a TryRemoveNode function that tries to remove a node from the graph. It does a DFS search after removing the node, and if the graph can be fully traversed, the node can be removed. If parts of the graph can not be reached the the removal will