Fixed Memory Leak
Fixed the memory leak when making large (and small, but didnt cause crash then) dungeons.
Issue was the visualization during the graph creation. This caused a copy being made for every room added to the graph. This would heap up and lead to accelerated memory allocation.

Added the visualization to the if block of the delay settings.