Refactor code for accesability
Refactored the delay functionality to be publicly available to all methods. The methods Delay() and DoInstantPass() are now in AlgorithmsUtils.cs.

Made the DungeonDataGenerator a singleton pattern so that the generationsettings are easily available. You can access them through the Instance. Otherwise i would have had to do Editor shenanigans and no ones likes that...